+++ internal/model/transaction.go
@@ -3,9 +3,9 @@ package model
 import "time"
 
 type Transaction struct {
-	ID        uint `gorm:"primary_key" json:"id"`
-	Sum       int  `gorm:"not null" json:"sum"`
-	UserID    uint `gorm:"not null" json:"userID"`
-	BookID    uint `gorm:"not null" json:"bookID"`
+	ID        uint    `gorm:"primary_key" json:"id"`
+	Sum       float64 `gorm:"not null" json:"sum"`
+	UserID    uint    `gorm:"not null" json:"userID"`
+	BookID    uint    `gorm:"not null" json:"bookID"`
 	CreatedAt time.Time
 }
